Feat: formatted firebase coordinate data into data readable by deck.gl, got an overlay to render
