--- client/data.js
@@ -1,108 +1,0 @@
-/// app.js
-import React, {useState} from 'react'
-import DeckGL from '@deck.gl/react'
-import {
-  StaticMap,
-  MapContext,
-  NavigationControl,
-  GeolocateControl,
-  Source,
-  Layer
-} from 'react-map-gl'
-import {SolidPolygonLayer} from '@deck.gl/layers'
-import {style} from './style'
-import {info} from './coordinates'
-
-// Set your mapbox access token here
-const MAPBOX_ACCESS_TOKEN =
-  'pk.eyJ1Ijoia2VuZGltb3Jhc2tpIiwiYSI6ImNra2U4YmpnODA4bXIycHA3dnA3ZHRxazMifQ.Xj6bAzbzUVih02szrFGa_Q'
-
-// Viewport settings
-const INITIAL_VIEW_STATE = {
-  longitude: -122.41669,
-  latitude: 37.7853,
-  zoom: 8,
-  pitch: 0,
-  bearing: 0
-}
-
-const geolocateControlStyle = {
-  // right: 10,
-  // top: 10
-}
-
-const MAP_STYLE = style
-
-const NAV_CONTROL_STYLE = {
-  position: 'absolute',
-  top: 10,
-  left: 10
-}
-
-const Data = () => {
-  const [viewport, setViewport] = useState({
-    latitude: 44.952261122619916,
-    longitude: -93.29339647810357,
-    width: '100wh',
-    height: '100vh',
-    zoom: 2
-  })
-  const [selectAdminLines, setAdminLines] = useState(false)
-  const layer = new SolidPolygonLayer({
-    data: info,
-    opacity: 0.5,
-    getPolygon: d => d.polygon,
-    getFillColor: [50, 147, 111],
-    extruded: false
-  })
-  return (
-    <DeckGL
-      initialViewState={INITIAL_VIEW_STATE}
-      controller={true}
-      layers={layer}
-      ContextProvider={MapContext.Provider}
-    >
-      <StaticMap
-        mapStyle={MAP_STYLE}
-        mapboxApiAccessToken={MAPBOX_ACCESS_TOKEN}
-      />
-      <NavigationControl style={NAV_CONTROL_STYLE} />
-      <GeolocateControl
-        style={geolocateControlStyle}
-        positionOptions={{enableHighAccuracy: true}}
-        trackUserLocation={true}
-        auto
-      />
-      <label
-        onClick={evt => {
-          setAdminLines(!selectAdminLines)
-        }}
-        className="adminContainer"
-      >
-        Admin Lines
-        <input type="checkbox" />
-        <span className="checkmark" />
-      </label>
-      {selectAdminLines ? (
-        <Source
-          id="adminLines"
-          type="vector"
-          url="mapbox://mapbox.mapbox-streets-v8"
-        >
-          <Layer
-            id="adminLines"
-            type="line"
-            source="admin-1"
-            source-layer="admin"
-            paint={{
-              'line-color': '#CAB69E',
-              'line-width': 0.75
-            }}
-          />
-        </Source>
-      ) : null}
-    </DeckGL>
-  )
-}
-
-export default Data
