+++ package.json
@@ -36,6 +36,7 @@
   "author": "Fullstack Academy of Code",
   "license": "MIT",
   "dependencies": {
+    "@babel/runtime-corejs3": "^7.13.10",
     "@firebase/firestore": "^2.2.1",
     "@google-cloud/firestore": "^4.9.8",
     "axios": "^0.19.0",
@@ -61,6 +62,7 @@
     "redux": "^4.0.1",
     "redux-logger": "^3.0.6",
     "redux-thunk": "^2.3.0",
+    "semver": "^6.3.0",
     "sequelize": "^5.3.1",
     "socket.io": "^2.2.0"
   },
